Move DistributedContext implementation into api level.
